third_party/WebKit/LayoutTests/external/wpt/fetch/api/cors/cors-expose-star-worker-expected.txt

This is a testharness.js-based test.
FAIL Basic Access-Control-Expose-Headers: * supportassert_equals: expected (string) "X" but got (object) null
PASS Cannot use * for credentialed fetches
Harness: the test ran to completion.
